fringes/requirements.txt
pillow
numpy
matplotlib
scipy
ipympl